improved ELF_Loader;+display symbols;+precompile errors without valid target

+++ src/kernel/HAL/x86/cmos.h
@@ -30,8 +30,8 @@
 static inline uint8_t cmos_read(uint8_t off){
 	uint8_t tmp =0;
 	uint8_t data =0;
-	INB(CMOS_PORT_ADDRESS,tmp)
-	OUTB(CMOS_PORT_ADDRESS,(tmp & 0x80) | (off & 0x7F))
+	INB((uint16_t)CMOS_PORT_ADDRESS,tmp)
+	OUTB((uint16_t)CMOS_PORT_ADDRESS,(tmp & 0x80) | (off & 0x7F))
 	INB(CMOS_PORT_DATA,data)
 	return data;
 }
